subiendo proyecto de register secutiry para exponer las funcionalidades de registro de uausior y clientes

+++ billers-commons/src/main/java/mx/budgie/commons/utils/CommonsUtil.java
@@ -89,7 +89,7 @@ public class CommonsUtil {
 		return differenceTime;
 	}
 	
-	public static Map<String, String> transactionIdHeader(final int systemId) {
+	public static Map<String, String> transactionIdHeader(final Long systemId) {
 		Map<String, String> headers = new LinkedHashMap<>();
 		headers.put("transactionId", String.valueOf(CommonsUtil.getTransactionId(systemId)));
 		return headers;
